Initial lift to develop an alternate strategy to compensate for the loss of `Transition::from`.
The idea is to give ownership of the most recent state to the learner, and have the learner cache that most recent state. That way at each learning step the learner has access both to the prior state and the most recent.

This isn't compiling yet. An initial, naive take is included, but now we need to switch around ownership since there were some conflicts there.

If we're clever, we should be able to free domains to emit only one clone of their state at each iteration.

In the future, we may be able to transition to a clone-free learning paradigm. This might be accomplished if, rather than caching a copy of the state, the learners cache only those functions of state which they would compute on its basis. Need to look more into this.